pyproject.toml
[tool.poetry]
name = "lombok-plugin-repository"
version = "1.0.0-alpha01"
description = "A repository for Lombok plugin incompatibility issues with Android Studio."
authors = ["Madray Haven <sgpublic2002@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "lombok_plugin_repository"}]

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0"
argparse = "^1.4.0"
loguru = "^0.7.0"
attrs = "^23.1.0"
apscheduler = "^3.10.1"
python-crontab = "^2.7.1"
gitpython = "^3.1.31"
croniter = "^1.3.15"
requests = "^2.31.0"
sortedcontainers = "^2.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
